+++ Sources/KeyboardKit/Actions/KeyboardAction+Autocomplete.swift
@@ -17,9 +17,7 @@ public extension KeyboardAction {
     var shouldApplyAutocompleteSuggestion: Bool {
         switch self {
         case .character(let char): return char.isWordDelimiter
-        case .newLine: return true
         case .primary(let type): return type.isSystemAction
-        case .return: return true
         case .space: return true
         default: return false
         }