+++ Sources/KeyboardKit/_Deprecated/Gestures+Deprecated.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-#if os(iOS) || os(tvOS)
-public extension SpaceCursorDragGestureHandler {
-
-    @available(*, deprecated, renamed: "keyboardContext")
-    var context: KeyboardContext { keyboardContext }
-}
-#endif