+++ Tests/KeyboardKitTests/Features/FeatureToggleTests.swift
@@ -17,16 +17,5 @@ final class FeatureToggleTests: XCTestCase {
         toggle = .shared
     }
 
-    func result(for feature: FeatureToggle.Feature) -> Bool {
-        toggle.isFeatureEnabled(feature)
-    }
-
-    func testCanToggleNewButtonGestureEngine() {
-        toggle.toggleFeature(.placeholder, .off)
-        XCTAssertFalse(result(for: .placeholder))
-        toggle.toggleFeature(.placeholder, .on)
-        XCTAssertTrue(result(for: .placeholder))
-        toggle.reset()
-        XCTAssertFalse(result(for: .placeholder))
-    }
+    func testCanToggleFeature() {}
 }