+++ RELEASE_NOTES.md
@@ -38,6 +38,8 @@ Some things that are planned for the 7.0 release:
 ### ✨ New features
 
 * `KeyboardInputViewController` has a new `viewWillSyncWithContext` function.
+* `KeyboardInputViewController` has a new `keyboardTextContext` property.
+* `KeyboardInputViewController` injects its `keyboardTextContext` into the environment.
 * `KeyboardTextContext` is a new observable class that lets you observe the document text.
 
 ### 👑 KeyboardKit Pro